fixing file name and moving extension methods
