+++ Assets/Mirage/Runtime/ClientStoppedReason.cs
@@ -1,4 +1,5 @@
-﻿using System;
+using System;
+using Mirage.SocketLayer;
 
 namespace Mirage
 {
@@ -30,5 +31,33 @@ namespace Mirage
         ConnectingTimeout = 5,
         /// <summary>Disconnect called locally before server replies with connected</summary>
         ConnectingCancel = 6,
+    }
+
+
+    internal static class StoppedReasonExtensions
+    {
+        public static ClientStoppedReason ToClientStoppedReason(this DisconnectReason reason)
+        {
+            switch (reason)
+            {
+                default:
+                case DisconnectReason.None: return ClientStoppedReason.None;
+                case DisconnectReason.Timeout: return ClientStoppedReason.Timeout;
+                case DisconnectReason.RequestedByRemotePeer: return ClientStoppedReason.RemoteConnectionClosed;
+                case DisconnectReason.RequestedByLocalPeer: return ClientStoppedReason.LocalConnectionClosed;
+            }
+        }
+
+        public static ClientStoppedReason ToClientStoppedReason(this RejectReason reason)
+        {
+            switch (reason)
+            {
+                default:
+                case RejectReason.None: return ClientStoppedReason.None;
+                case RejectReason.Timeout: return ClientStoppedReason.ConnectingTimeout;
+                case RejectReason.ServerFull: return ClientStoppedReason.ServerFull;
+                case RejectReason.ClosedByPeer: return ClientStoppedReason.ConnectingCancel;
+            }
+        }
     }
 }
